backend/requirements.txt
flask
db-sqlite3
google-cloud-aiplatform